[MOVE] article-refactoring / #71
article 관련 뷰 패키지 분리
