+++ app/src/main/java/com/indipage/presentation/articledetail/ArticleDetailViewModel.kt
@@ -1,4 +1,4 @@
-package com.indipage.presentation.article
+package com.indipage.presentation.articledetail
 
 import androidx.lifecycle.LiveData
 import androidx.lifecycle.MutableLiveData
