+++ app/src/main/res/navigation/navigation_main.xml
@@ -74,7 +74,7 @@
     
     <fragment
         android:id="@+id/navigation_article_detail"
-        android:name="com.indipage.presentation.article.ArticleDetailFragment"
+        android:name="com.indipage.presentation.articledetail.ArticleDetailFragment"
         android:label="ArticleDetailFragment" >
         <action
             android:id="@+id/action_article_detail_to_space_detail"
@@ -93,7 +93,7 @@
     </fragment>
     <fragment
         android:id="@+id/navigation_article_all"
-        android:name="com.indipage.presentation.article.ArticleAllFragment"
+        android:name="com.indipage.presentation.articleall.ArticleAllFragment"
         android:label="ArticleAllFragment" >
         <action
             android:id="@+id/action_article_all_to_article"
